metadata type 33 is not defined
metadata type 33 or "Midi Port" acording to this website -> https://mido.readthedocs.io/en/1.1.24/meta_message_types.html <- was not defined in Constants.METADATA_TYPES.
Loading a midi file that uses the metadata type 33 causes the program to fail with the error "Table index is nil".
Is there a reason why these other types are not defined?